+++ src/XIVLauncher.Common.Unix/UnixDalamudRunner.cs
@@ -44,6 +44,8 @@ public class UnixDalamudRunner : IDalamudRunner
 
         environment.Add("DALAMUD_RUNTIME", dotnetRuntimePath);
         environment.Add("DOTNET_ROOT", dotnetRuntimePath);
+        // Fix for GE-Wayland patches after Nov 26 2025.
+        environment.Add("DOTNET_ROOT_X64", dotnetRuntimePath);
 
         var launchArguments = new List<string>
         {